[pulsar-broker]Cleanup all metadata inited by initialize-cluster-metadata. (#9474)

+++ pulsar-broker/src/main/java/org/apache/pulsar/PulsarClusterMetadataTeardown.java
@@ -30,6 +30,7 @@ import org.apache.bookkeeper.conf.ClientConfiguration;
 import org.apache.bookkeeper.mledger.ManagedLedgerException;
 import org.apache.bookkeeper.mledger.ManagedLedgerFactory;
 import org.apache.bookkeeper.mledger.impl.ManagedLedgerFactoryImpl;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.pulsar.broker.service.schema.SchemaStorageFormat.SchemaLocator;
 import org.apache.pulsar.common.naming.TopicName;
 import org.apache.pulsar.zookeeper.ZooKeeperClientFactory;
@@ -55,21 +56,32 @@ public class PulsarClusterMetadataTeardown {
         }, description = "Local zookeeper session timeout ms")
         private int zkSessionTimeoutMillis = 30000;
 
-        @Parameter(names = { "-c", "-cluster" }, description = "Cluster name")
+        @Parameter(names = { "-c", "--cluster" }, description = "Cluster name")
         private String cluster;
 
         @Parameter(names = { "-cs", "--configuration-store" }, description = "Configuration Store connection string")
         private String configurationStore;
 
-        @Parameter(names = { "--bookkeeper-metadata-service-uri" }, description = "Metadata service uri of BookKeeper")
+        // Hide and marked as deprecated this flag because we use the new name '--existing-bk-metadata-service-uri' to
+        // pass the service url. For compatibility of the command, we should keep both to avoid the exceptions.
+        @Deprecated
+        @Parameter(names = {
+                "--bookkeeper-metadata-service-uri"},
+                description = "The metadata service URI of the existing BookKeeper cluster that you used",
+                hidden = true)
         private String bkMetadataServiceUri;
+
+        @Parameter(names = {
+                "--existing-bk-metadata-service-uri"},
+                description = "The metadata service URI of the existing BookKeeper cluster that you used")
+        private String existingBkMetadataServiceUri;
 
         @Parameter(names = { "-h", "--help" }, description = "Show this help message")
         private boolean help = false;
     }
 
     public static String[] localZkNodes = {
-            "bookies", "counters", "loadbalance", "managed-ledgers", "namespace", "schemas", "stream" };
+            "bookies", "admin", "ledgers", "managed-ledgers", "namespace", "pulsar", "stream" };
 
     public static void main(String[] args) throws Exception {
         Arguments arguments = new Arguments();
@@ -86,9 +98,12 @@ public class PulsarClusterMetadataTeardown {
             throw e;
         }
 
-        if (arguments.bkMetadataServiceUri != null) {
-            BookKeeper bookKeeper =
-                    new BookKeeper(new ClientConfiguration().setMetadataServiceUri(arguments.bkMetadataServiceUri));
+        if (arguments.bkMetadataServiceUri != null || arguments.existingBkMetadataServiceUri != null) {
+            String uri = arguments.existingBkMetadataServiceUri;
+            if (uri == null){
+                uri = arguments.bkMetadataServiceUri;
+            }
+            BookKeeper bookKeeper = new BookKeeper(new ClientConfiguration().setMetadataServiceUri(uri));
             ZooKeeper localZk = initZk(arguments.zookeeper, arguments.zkSessionTimeoutMillis);
             ManagedLedgerFactory managedLedgerFactory = new ManagedLedgerFactoryImpl(bookKeeper, localZk);
 
@@ -102,6 +117,10 @@ public class PulsarClusterMetadataTeardown {
         ZooKeeper localZk = initZk(arguments.zookeeper, arguments.zkSessionTimeoutMillis);
 
         for (String localZkNode : localZkNodes) {
+            // when use --existing-bk-metadata-service-uri doesn't contain "/ledgers"
+            if (arguments.existingBkMetadataServiceUri != null && StringUtils.equals("ledgers", localZkNode)) {
+                continue;
+            }
             deleteZkNodeRecursively(localZk, "/" + localZkNode);
         }
